moved sign out to nav bar
